feat/157 leave button
