+++ frontend/src/Chat/components/ChatRoom/UserListDisplay.tsx
@@ -1,34 +1,10 @@
 import { type ReactElement } from 'react'
-import { Button } from 'react-bootstrap'
 import { OwnerIcon } from '../utils/OwnerIcon'
 import { AdminButton } from './AdminButton'
 import { BanButton } from './BanButton'
-import { deleteChatRoomMembersRequest } from '../utils/requestUtils'
-import { isAdmin, isOwner } from '../utils/userStatusUtils'
+import { isAdmin } from '../utils/userStatusUtils'
 import { AdminIcon } from '../utils/AdminIcon'
-
-const DeleteMemberButton = (props: {
-  room: ChatRoom
-  member: User
-  updateMemberList: () => void
-}): JSX.Element => {
-  if (isOwner(props.member, props.room)) return <></>
-
-  return (
-    <Button
-      variant="outline-danger"
-      onClick={() => {
-        const requestData: ChatRoomMemberPK = {
-          chatRoomId: props.room.id,
-          userId: props.member.id
-        }
-        deleteChatRoomMembersRequest(requestData, props.updateMemberList)
-      }}
-    >
-      Delete User from Room
-    </Button>
-  )
-}
+import { DeleteMemberButton } from '../utils/DeleteMemberButton'
 
 export const UserListDisplay = (props: {
   room: ChatRoom
@@ -49,7 +25,12 @@ export const UserListDisplay = (props: {
             />
             <BanButton {...props} member={member} />
             <AdminButton {...props} member={member} />
-            <DeleteMemberButton {...props} member={member} />
+            <DeleteMemberButton
+              {...props}
+              member={member}
+              onClickCallback={props.updateMemberList}
+              msg={'Delete User from Room'}
+            />
           </li>
         )
       })}
